### 2025-03-08 #234 @c4dynamics
C4dynamics submission
Submitting Author: Ziv Meri (@C4dynamics)
All current maintainers: Ziv Meri (@C4dynamics)
Package Name: C4dynamics
One-Line Description of Package: Python framework for algorithms of dynamic systems
Repository Link:  https://github.com/C4dynamics/C4dynamics
Version submitted: v2.0.0
EiC: @coatless, @crhea93, @yeelauren
Version accepted: TBD

---

c4dynamics is designed to simplify the development of algorithms for dynamic systems, using state space representations. It offers engineers and researchers a systematic approach to model, simulate, and control systems in fields like robotics, aerospace, and navigation.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

c4dynamics belongs to the category of math-operations and data-operations for physical models. The different data processing/munging operations are overviewed in the documentary page of the package core: [state operations] (https://c4dynamics.github.io/C4dynamics/api/States.html#operations)

Researchers, engineers, and students from the fields of aerospace, robotics, navigation guidance and control.

For event-driven or discrete event simulation, [statepy](https://pypi.org/project/StatePy/) [defunct], [statemachine](https://python-statemachine.readthedocs.io/en/latest/readme.html), or [transition](https://github.com/pytransitions/transitions).
For decision-making or reinforcement learning systems, [POMDPy](https://github.com/pemami4911/POMDPy) (not maintained) and tools like transition (for FSM).

https://github.com/pyOpenSci/software-submission/issues/224#issuecomment-2703355782

Looking forward working with you. Thanks!
