### 2025-05-07 #234 @crhea93
---
---

The authors of this package provide very detailed documentation in the form of detailed API documentation and 3 detailed tutorials. They also have a simple to follow quick start. My only suggestion in terms of the documentation is that a lot of the really well-documented functionality is buried in links throughout the documentation. I'd suggest rethinking how the left-hand navigation panel could be redone to make it easier to access the documentation. For example, there is a really well fleshed-out page on the Rigidbody state, but, at a first glance of the documentation, it is not clear how to get there (I had to go through a few other pieces to end up there). I'd like to stress, however, that the documentation is very extensive!

A final note on documentation: I'd love to see a section breaking down the math/physics of the main concepts. Personally, that might help a user understand where to get started.

The ReadME file is very comprehensive until we reach the section entitled Block 2. It looks like the author was in the process of completing the ReadME file but did not finish this section.

The unit tests are passing (this is clear from the Github Action and the button at the top of the ReadME). However, it would be nice to see the coverage. I note that there is something happening with codedev in the Action...
