### 2025-05-16 #234 @c4dynamics
Hi @crhea93  👋

Thank you very much for your professional review!

Here are my responses to two of your four points:

1. README
I've updated the "Block 2" section to:
"[New in Block 2](https://github.com/C4dynamics/C4dynamics?tab=readme-ov-file#new-in-block-2)
Enhancements and modules in latest release:... "
Let me know if that now reads smoothly and looks complete.

2. Documentation
Regarding your feedback on navigation and clarity:
I'm planning to add a new section called "Concepts" or "Fundamentals" before the API reference. This would centralize the key mathematical and physical ideas (like state vectors, transformations, filters, etc.) that are currently scattered across the docs.
The idea is similar to [NumPy’s Fundamentals](https://numpy.org/doc/2.2/user/basics.html), with conceptual overviews and links to the relevant API pages.
Do you think this structure would help? Or maybe you have any favorite documentation site I could look at for inspiration?

Thanks again for the thoughtful feedback. 🙏
