### 2025-05-16 #234 @crhea93
@C4dynamics
Thank you for addressing those two points! The README is perfect now :)

As for the documentation, I think that would be excellent. That would certainly help readability and show off your excellent documentation.
