### 2025-09-30 #234 @c4dynamics
Hi @crhea93 ,

I hope you’re doing well! I wanted to let you know that I’ve completed the documentation updates we discussed here.
Apologies for the delay — the changes were quite substantial, including the new “Concepts” section summarizing state vectors, transformations, and filters, and a reorganization of the navigation to make the content more accessible.

I’d be happy to hear any feedback or suggestions on whether this aligns with your recommendations.
https://c4dynamics.github.io/C4dynamics

Thanks so much for your time!

Ziv
