### 2025-12-13 #234 @c4dynamics
Hey @yeelauren, hope you're doing well!
Just checking - no rush - any update on the search for editors?
Thanks, Ziv.
