### 2025-12-22 #234 @crhea93
Hi @C4dynamics,
 I'm very pleased to be working as editor for your submission! I have a few contacts I'll be reaching out to this week who I think would be excellent reviewers for your package; I'll also be reaching out to the community at PyOpenSci in general to serach for additional potential reviewers. However, since we are entering the holidays, I wanted to let you know that it may take a little longer than usual to obtain reviewers/for the reviews to begin. I greatly appreciate your patience with this submission!
