### 2026-01-15 #234 @crhea93
## Editor response to review:

---

## Editor comments
Thank you very much for your patience at @c4dynamics! I've found two excellent reviewers with, I think, the perfect mix of backgrounds for this review.

:wave: Hi @adityapt and @hfactor13! Thank you for volunteering to review for pyOpenSci!

Please get in touch with any questions or concerns! Your review is due: February 10th, 2026
---

Reviewers: @hfactor13 @adityapt
Due date: 2026-02-10
