### 2026-01-24 #234 @hfactor13
Here's my review. I am still trying to troubleshoot the unittests that are failing on my machine.

- [X] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [X] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file
- [X] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [X] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

  - [X] Continuous integration and test coverage
  - [X] Docs building (if you have a documentation website)
  - [X] Python versions supported
  - [ ] Current package version (on PyPI / Conda)

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

The package structure should follow the general community best practices. In general, please consider whether:

8 hrs

##### Documentation Comments

###### Statement of Need

I would move the checklist to the bottom and have a separate section called about c4dynamics. [Section Rearrangement of README.md](https://github.com/c4dynamics/c4dynamics/issues/137)

###### Installation Instructions

I would add an installation instruction for the `conda` package. Maybe something along the lines of `conda install -c conda-forge c4dynamics`. I think it's worth having if users are working with `conda` and don't have too much familiarity with the different package management systems. I know when I started in my Python journey, I used the Anaconda Navigator a lot and didn't really use `pip` until I was contributing to another person's Python codebase. [Add Conda Installation Instructions](https://github.com/c4dynamics/c4dynamics/issues/140)

###### Short Quickstart Tutorials

When I ran the quick start tutorial, I noticed that on the last piece of code denoted `s.data('y')`, I only received the second list item `[1, 2.5]`, but in the example, it shows a tuple of two lists `([0, 1], [1, 2.5])`. [Short Tutorial](https://github.com/c4dynamics/c4dynamics/issues/139)

###### Community Guidelines

I noticed that there is a broken link to the `Use Cases` section. Maybe just route the link to the documentation website. [Fix Broken Hyperlink](https://github.com/c4dynamics/c4dynamics/issues/138)

##### README File Comments

###### Badges

There's a version number for PyPI, but not Conda.

###### Additional Setup

I think the `pooch` module should be added as a dependency when installing `c4dynamics`. I was using Python version 3.12, and I got an error message that I needed to install `pooch`. The attachment shows the packages that are installed after running `pip install c4dynamics`.  [Missing Dependency](https://github.com/c4dynamics/c4dynamics/issues/141)

###### Vignettes

I personally didn't see any, but maybe I may have missed it somewhere.

##### Functionality Comments

###### Automated Tests

Based on the testing instructions provided in the `contributing.md` file, it's not clear what the test coverage is (I am looking for a percentage in the CLI output). See the example attachment below. I noticed that the standard built-in `unittest` is used in the repo among your testing modules. I am more familiar with the `pytest` framework, and I am not sure if there's a `unittest` equivalent to `pytest-cov`. I believe the `coverage` package in PyPI can still work with `unittest`, but I have never personally tried it. Either way, I would recommend adding a package that can do the test coverage and have a way to filter out passing tests so that we only see what really fails. Currently, I am trying to identify which tests have failed and need attention amid a sea of passed tests. At a glance, I did notice some tests fail due to `NumPy` and `PyQt`. [Improve Testing Coverage](https://github.com/c4dynamics/c4dynamics/issues/142)

###### Packaging Guidelines

Based on the end-of-life website, 3.9 shouldn't be supported anymore. Need to stick to version number 3.10 and above. I didn't see in the CI anything about checking the formatting.
