### 2026-01-25 #234 @c4dynamics
Hi @hfactor13,
Thanks for the thoughtful review!
I’ll go through it carefully and act on it ASAP.
