### 2026-01-25 #234 @adityapt
I will be starting my review here.

- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [x] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [x] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

The package structure should follow the general community best practices. In general, please consider whether:

---

1) The installation fails on version 3.12 due to OpenCV in certain environments. The Kernel completely crashes.
2) The ReadMe file doesn't mention the list of python versions this package works perfectly for.
3) The package is not so easy to install. One or the other dependency issue comes up. On DBx it was OpenCV, and on my local, it was tkinker.
4) The package seems to not work on Python 3.13, so I initially tested it with old version. Upon creating a new Venv with 3.12 and using 2.3.3, most tests pass, with some errors.

Ran tests on my local computer:

Passing Tests (20 files):

test_gen_gif.py
test_animate.py (6 skipped)
test_datasets.py
test_lowpass.py
test_tictoc.py
test_const.py
test_state.py  (17 tests)
test_pixelpoint.py
test_yolov3_opencv.py
test_cprint.py
test_kalman.py
test_derivs.py
test_ekf.py
test_integrate.py
test_datapoint.py
test_radar.py
test_seeker.py
test_math.py
test_rigidbody.py
test_rotmat.py (1 skipped)

Minor Issues (3 files):
test_state_assignment.py - AssertionError
Expected integer dtype, got float64

test_state_assignment_strict.py - AssertionError
Expected integer dtype initially

test_plotdefaults.py - 1 FAILURE
DPI mismatch: got 400.0, expected 200

The test coverage is good, but not exhaustive.
