### 2026-03-21 #234 @c4dynamics
Hi @crhea93,

Thanks for following up.

I’ve addressed all the comments from @adityapt.
At this point, the only remaining item is from @hfactor13 regarding adding support for conda.

I’ve already opened [a PR for this](https://github.com/conda-forge/staged-recipes/pull/32341) about three weeks ago.

However, I haven’t received a response there yet.
Once that’s resolved, everything should be complete.

Please let me know if there’s anything else I should do in the meantime.

Best regards,
