### 2026-04-12 #234 @hfactor13
@crhea93 Everything is good on my end. I'm just waiting on the PR for the `conda` registration of the package. Here is the [issue number](https://github.com/c4dynamics/c4dynamics/issues/140) and corresponding [PR](https://github.com/conda-forge/staged-recipes/pull/32341).
