### 2026-04-13 #234 @c4dynamics
Hi @crhea93, @hfactor13,

Thanks for the update.

Since this is the last remaining step on my side and everything else is complete, I wanted to ask whether it would be possible to proceed without waiting for the merge, or if the conda-forge publication is strictly required before completion.

Happy to make any additional adjustments if needed.
Thanks again for your guidance.

Ziv.
